Adopt a staged Clippy lint policy and enforce three tranches (#74)
Move the Clippy configuration out of the long `-A clippy::...` lists in
windows-ci.yml and release.yml into `[workspace.lints.clippy]` (Cargo.toml);
each crate opts in via `[lints] workspace = true`. Both CI steps are now
simply `cargo clippy --workspace --all-targets -- -D warnings`.

Enforce three tranches of previously-suppressed lints, fixing genuine
findings and adding narrowly-scoped `#[allow]`s (with rationale) only
where a warning is intentional:

- Tranche 1: large_enum_variant, len_without_is_empty, unnecessary_unwrap,
  iter_overeager_cloned, manual_contains, result_large_err.
- Tranche 2: derivable_impls, unnecessary_cast, needless_borrow,
  approx_constant, len_zero, needless_return, redundant_closure,
  useless_conversion, question_mark, unnecessary_lazy_evaluations,
  single_match, needless_update.
- Tranche 3: collapsible_if, collapsible_match,
  empty_line_after_doc_comments, items_after_test_module,
  vec_init_then_push, unusual_byte_groupings, excessive_precision,
  ptr_arg, needless_lifetimes, default_constructed_unit_structs,
  cloned_ref_to_slice_refs, disallowed_names, assertions_on_constants,
  if_same_then_else, identity_op.

The remaining allowlist contains only lints deferred for a documented
reason: behavior-risk (manual_clamp), large churn (manual_div_ceil,
field_reassign_with_default), explicitly out of scope per #74
(too_many_arguments, needless_range_loop), case-by-case judgment
(let_unit_value, type_complexity, needless_bool_assign), and
absurd_extreme_comparisons pending the paint-debounce cleanup.

Refs #74.

+++ crates/terra-core/src/landscape_evolution/iterative.rs
@@ -152,7 +152,7 @@ fn slope_to_receiver(z: &Heightfield, i: usize, j: usize, cache: &DrainageCache)
         let di = ri as i32 - i as i32;
         let dj = rj as i32 - j as i32;
         let dist = if di != 0 && dj != 0 {
-            ((dx * dx + dz * dz) as f32).sqrt()
+            (dx * dx + dz * dz).sqrt()
         } else if di != 0 {
             dx
         } else {
@@ -169,7 +169,7 @@ fn slope_to_receiver(z: &Heightfield, i: usize, j: usize, cache: &DrainageCache)
             continue;
         }
         let dist = if di != 0 && dj != 0 {
-            ((dx * dx + dz * dz) as f32).sqrt()
+            (dx * dx + dz * dz).sqrt()
         } else if di != 0 {
             dx
         } else {
